Server/requirements.txt
fastapi
pandas
scikit-learn
joblib
python-whois
numpy
uvicorn
